fix: typo in docs and remove unnecessary calls to Uint64 function

+++ examples/scratch-storage/contract.algo.ts
@@ -7,7 +7,7 @@ export class ScratchSlotsContract extends arc4.Contract {
   public storeData(): boolean {
     op.Scratch.store(1, Uint64(5))
     op.Scratch.store(2, Bytes('Hello World'))
-    assert(op.Scratch.loadUint64(1) === Uint64(5))
+    assert(op.Scratch.loadUint64(1) === 5)
     assert(op.Scratch.loadBytes(2) === Bytes('Hello World'))
     return true
   }
@@ -16,7 +16,7 @@ export class ScratchSlotsContract extends arc4.Contract {
 @contract({ scratchSlots: [1, 2] })
 export class SimpleScratchSlotsContract extends BaseContract {
   approvalProgram(): uint64 {
-    assert(op.Scratch.loadUint64(1) === Uint64(5))
+    assert(op.Scratch.loadUint64(1) === 5)
     assert(op.Scratch.loadBytes(2) === Bytes('Hello World'))
     return Uint64(1)
   }
